+++ src/decoder/aac_decoder.ml
@@ -35,7 +35,7 @@ exception End_of_stream
 (** Buffered input device where
   * the buffer initially contains [Bytes.sub buf offset len]. *)
 let buffered_input input buf offset len =
-  let buffer = Buffer.create 1024 in
+  let buffer = Buffer.create Utils.pagesize in
   let pos = ref len in
   Buffer.add_subbytes buffer buf offset len;
   let drop len = 
@@ -60,7 +60,7 @@ let buffered_input input buf offset len =
   (* Get at most [len] bytes from the buffer,
    * which is refilled from [input] if needed.
    * This does not remove data from the buffer. *)
-  let tmplen = 1024 in
+  let tmplen = Utils.pagesize in
   let tmp = Bytes.create tmplen in
   let read buf ofs len =
     let size = Buffer.length buffer in
@@ -88,8 +88,7 @@ struct
 let create_decoder input =
   let resampler = Rutils.create_audio () in
   let dec = Faad.create () in
-  (* 1024 bytes seems usually enough to initiate the decoder.. *)
-  let initbuflen = 1024 in
+  let initbuflen = Utils.pagesize in
   let initbuf = Bytes.create initbuflen in
   let len = input.Decoder.read initbuf 0 initbuflen in
   let offset, sample_freq, chans =
@@ -167,7 +166,7 @@ let get_type filename =
   Tutils.finalize ~k:(fun () -> Unix.close fd)
     (fun () ->
       let dec = Faad.create () in
-      let aacbuflen = 1024 in
+      let aacbuflen = Utils.pagesize in
       let aacbuf = Bytes.create aacbuflen in
       let _,rate,channels =
         let n = Unix.read fd aacbuf 0 aacbuflen in